Move `TranslateFluent` implementations to base class
